backend/requirements.txt
# Backend deployment requirements
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
slowapi>=0.1.9

# AI APIs
openai>=1.0.0
google-genai>=0.8.0

# Data processing
numpy>=1.24.0

# HTTP requests
requests>=2.31.0

# Production server
gunicorn>=21.2.0

# Optional caching
redis>=5.0.0

# Optional vector index
faiss-cpu>=1.8.0
